stats: add a simple bar plot and wire it to the frontend
The most annoying part is actually not the plot itself,
but the whole combo-box back-and-forth to let the user
select different axis types and bins.

This doesn't yet do anything too useful. The user can now
plot dive type against buddy - in the unlikely case that they
ever wanted to do that. :-P

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>